Added some changes

+++ sort.h
@@ -75,5 +75,5 @@ void quick_sort_hoare(int *array, size_t size);
 void sort_deck(deck_node_t **deck);
 
 /*Utilities function prototypes*/
-void swap(int *array, int position);
+void swap(int *array, size_t position);
 #endif
